SolutionUI: Move all constants to a single file
