+++ examples/metalk8s-solution-example/ui/src/constants.js
@@ -1,1 +1,10 @@
 export const REFRESH_TIMEOUT = 15000;
+
+export const LABEL_COMPONENT = 'app.kubernetes.io/component';
+export const LABEL_NAME = 'app.kubernetes.io/name';
+export const LABEL_PART_OF = 'app.kubernetes.io/part-of';
+export const LABEL_VERSION = 'app.kubernetes.io/version';
+export const SOLUTION_NAME = 'example-solution';
+export const DEPLOYMENT_NAME = 'example-operator';
+export const OPERATOR_NAME = 'example-solution-operator';
+export const SOLUTION_CONFIGMAP_NAME = 'metalk8s-solutions';
